+++ src/yt_maint.h
@@ -313,7 +313,7 @@ bool yt_maintenance_xannor_target_override(int group_number,
     int headquarters, int revenge_live_sector, int top_player_target,
     int *target, struct yt_error *error);
 bool yt_maintenance_xannor_defense(struct yt_random *random,
-    float *group_size, float *defense_fighters, float *defense_owner,
+    float *group_size, float *defense_fighters, int *defense_owner,
     struct yt_error *error);
 bool yt_maintenance_xannor_sector_arrival(struct yt_game *game,
     int sector_number, float *group_size, struct yt_sector *sector,
